% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{chimeraviz-internals-fusionsToLinkData}
\alias{chimeraviz-internals-fusionsToLinkData}
\alias{.fusionsToLinkData}
\title{Create link data for RCircos from the given fusions.}
\usage{
.fusionsToLinkData(fusionList, minLinkWidth = 1, maxLinkWidt = 10)
}
\arguments{
\item{fusionList}{A list of Fusion objects.}

\item{minLinkWidth}{The minimum link line width. Default = 1}

\item{maxLinkWidt}{The maximum link line width. Default = 10}
}
\value{
A data frame with fusion link data compatible with RCircos::RCircos.Link.Plot()

# @examples # Apparently examples shouldn't be set on private functions
defuse833ke <- system.file("extdata", "defuse_833ke_results.filtered.tsv", package="chimeraviz")
fusions <- importDefuse(defuse833ke, "hg19", 3)
linkData <- chimeraviz::.fusionsToLinkData(fusions)
# This linkData can be used with RCircos::RCircos.Link.Plot()
}
\description{
This function takes a list of Fusion objects and creates a data frame in the
format that RCircos::RCircos.Link.Plot() expects for link data.
}
