% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circle.R
\name{chimeraviz-internals-scaleListToInterval}
\alias{chimeraviz-internals-scaleListToInterval}
\alias{.scale_list_to_interval}
\title{Scale a vector of numeric values to an interval.}
\usage{
.scale_list_to_interval(the_list, new_min, new_max)
}
\arguments{
\item{the_list}{A vector of numeric values.}

\item{new_min}{Minimum value for the new interval.}

\item{new_max}{Maximum value for the new interval.}
}
\value{
A data frame with fusion link data compatible with RCircos::RCircos.Link.Plot()

# @examples # Apparently examples shouldn't be set on private functions
list012 <- c(0,1,2)
.scale_list_to_interval(list012, 1, 3)
# [1] 1 2 3
}
\description{
This function takes a vector of numeric values as well as an interval
[new_min, new_max] that the numeric values will be scaled (normalized) to.
}
