% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimeraviz.R
\docType{package}
\name{chimeraviz}
\alias{chimeraviz}
\title{chimeraviz: A package for working with and visualizing fusion genes.}
\description{
chimeraviz manages data from fusion gene finders and provides useful
visualization tools.
}
