% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{createFusionReport}
\alias{createFusionReport}
\title{Create a Fusion Report}
\usage{
createFusionReport(fusions, outputFilename, quiet = TRUE)
}
\arguments{
\item{fusions}{A list of Fusion objects.}

\item{outputFilename}{Output html-file filename.}

\item{quiet}{Parameter passed to rmarkdown::render() to toggle its output.}
}
\value{
Creates a html report with an overplot and a sortable, searchable
table with the fusion data.
}
\description{
This function will create a html report with an overplot and a sortable,
searchable table with the fusion data.
}
\examples{
# Load data
defuse833ke <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuse833ke, "hg19", 3)
# Temporary file to store the report
outputFilename <- tempfile(
  pattern = "fusionReport",
  fileext = ".html",
  tmpdir = tempdir())
# Create report
createFusionReport(fusions, outputFilename)

}

