% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{downstreamPartnerGene}
\alias{downstreamPartnerGene}
\alias{downstreamPartnerGene,Fusion-method}
\alias{downstreamPartnerGene<-}
\alias{downstreamPartnerGene<-,Fusion-method}
\title{Get the downstream fusion partner gene}
\usage{
downstreamPartnerGene(x)

\S4method{downstreamPartnerGene}{Fusion}(x)

downstreamPartnerGene(object) <- value

\S4method{downstreamPartnerGene}{Fusion}(object) <- value
}
\arguments{
\item{x}{The Fusion object you wish to retrieve the downstream PartnerGene object for.}

\item{object}{The Fusion object you wish to set a new downstream PartnerGene object for.}

\item{value}{The new PartnerGene object.}
}
\value{
The downstream PartnerGene object.
}
\description{
This getter retrieves the downstream PartnerGene object.

This sets the downstream PartnerGene object of a Fusion object
}
\examples{
# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
fusion <- fusions[[1]]
# Get the downstream fusion partner gene
downstreamPartnerGene(fusion)

# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
fusion <- fusions[[1]]
# Set the downstream PartnerGene object to be the same as the upstream PartnerGene object
downstreamPartnerGene(fusion) <- upstreamPartnerGene(fusion)

}
