% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{fusionSplitReadsCount}
\alias{fusionSplitReadsCount}
\alias{fusionSplitReadsCount,Fusion-method}
\title{Get the split reads count from a Fusion object}
\usage{
fusionSplitReadsCount(x)

\S4method{fusionSplitReadsCount}{Fusion}(x)
}
\arguments{
\item{x}{The Fusion object you wish to retrieve the split reads count for.}
}
\value{
The Fusion split reads count.
}
\description{
This getter retrieves the split reads count from a Fusion object
}
\examples{
# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
fusion <- fusions[[1]]
# Get the split reads count
fusionSplitReadsCount(fusion)

}
