% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fusionToDataFrame}
\alias{fusionToDataFrame}
\title{Coerce Fusion object to data.frame}
\usage{
fusionToDataFrame(fusion)
}
\arguments{
\item{fusion}{The Fusion object to coerce.}
}
\value{
A data.frame with the fusion object.
}
\description{
This function is used in createFusionReport() to convert Fusion objects to a
data.frame-format.
}
\examples{
# Load data
defuse833ke <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuse833ke, "hg19", 1)
# Find the fusion object to create a data frame from
fusion <- getFusionById(fusions, 5267)
# Create the data frame
dfFusion <- fusionToDataFrame(fusion)

}
\seealso{
createFusionReport
}
