% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getEnsemblIds}
\alias{getEnsemblIds}
\title{Get ensembl ids for a fusion object}
\usage{
getEnsemblIds(fusion)
}
\arguments{
\item{fusion}{The Fusion object we want to get ensembl ids for.}
}
\value{
The Fusion object with Ensembl ids set.
}
\description{
This function will get the ensembl ids from the org.Hs.eg.db package given
the gene names of the fusion event.
}
\examples{
# Import the filtered defuse results
defuse833keFiltered <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuse833keFiltered, "hg19", 1)
# Get a specific fusion
fusion <- getFusionById(fusions, 5267)
# See the ensembl ids:
fusion@geneA@ensemblId
# [1] "ENSG00000180198"
fusion@geneB@ensemblId
# [1] "ENSG00000162639"
# Reset the fusion objects ensembl ids
fusion@geneA@ensemblId <- ""
fusion@geneB@ensemblId <- ""
# Get the ensembl ids
fusion <- getEnsemblIds(fusion)
# See that we now have the same ensembl ids again:
fusion@geneA@ensemblId
# [1] "ENSG00000180198"
fusion@geneB@ensemblId
# [1] "ENSG00000162639"

}

