% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFusionBy.R
\name{getFusionByChromosome}
\alias{getFusionByChromosome}
\title{Find fusions that involves genes in the given chromosome.}
\usage{
getFusionByChromosome(fusionList, chr)
}
\arguments{
\item{fusionList}{A list of Fusion objects.}

\item{chr}{The chromosome name we're looking for fusions in.}
}
\value{
A list of Fusion objects.
}
\description{
Helper function to retrieve the Fusion objects that involves genes in the
given chromosome name.
}
\examples{
defuse833ke <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuse833ke, "hg19", 1)
length(getFusionByChromosome(fusions, "chr1"))
# [1] 1

}
