% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFusionBy.R
\name{getFusionByGeneName}
\alias{getFusionByGeneName}
\title{Find fusions that includes the given gene.}
\usage{
getFusionByGeneName(fusionList, geneName)
}
\arguments{
\item{fusionList}{A list of Fusion objects.}

\item{geneName}{The gene name we're looking for.}
}
\value{
A list of Fusion objects.
}
\description{
Helper function to retrieve the Fusion objects that has geneName as one of
the partner genes.
}
\details{
Note: getFusionByGeneName(fusionList, "MT") will match both MT-ND5 and
MT-ND4.
}
\examples{
defuse833ke <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuse833ke, "hg19", 1)
length(getFusionByGeneName(fusions, "RCC1"))
# [1] 1

}

