% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_aeron.R
\name{import_aeron}
\alias{import_aeron}
\title{Import results from an Aeron run into a list of Fusion objects.}
\usage{
import_aeron(filename, genome_version, limit)
}
\arguments{
\item{filename}{Filename for the Aeron results.}

\item{genome_version}{Which genome was used in mapping (hg19, hg38, etc.).}

\item{limit}{A limit on how many lines to read.}
}
\value{
A list of Fusion objects.
}
\description{
A function that imports the results from an Aeron run into a list of Fusion
objects.
}
\examples{
aeronfile <- system.file(
  "extdata",
  "aeron_fusion_support.txt",
  package="chimeraviz")
fusions <- import_aeron(aeronfile, "hg19", 3)
# This should import a list of 3 fusions described in Fusion objects.
}
