% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fusion.R
\name{import_function_non_ucsc}
\alias{import_function_non_ucsc}
\title{Alternative import function for Gviz::AlignmentsTrack}
\usage{
import_function_non_ucsc(file, selection)
}
\arguments{
\item{file}{The bamfile.}

\item{selection}{Which regions to get from the bamfile.}
}
\value{
A GRanges object with coverage data for the selection.
}
\description{
This alternative import function for use with Gviz::AlignmentsTrack imports
a bamfile with non-UCSC chromosome names.
}
