% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFusionInspector.R
\name{import_fusion_inspector}
\alias{import_fusion_inspector}
\title{read the results from runnning FusionInspector}
\usage{

  import_fusion_inspector(filename = "FusionInspector-inspect/finspector.igv.FusionJuncSpan")
}
\arguments{
\item{filename}{Which file to import. Note that}
}
\description{
This function is called when \code{\link{import_star_fusion}} is called
with \code{use_fusion_inspector=TRUE}. It will override coordinates
(chromosome, strand, and breakpoint) with those corresponding to
FusionInspector's "mini genome". The resulting list of fusion objects
can currently only be plotted with \code{\link{plotFusionReadsSimple}}.
}
\details{
Note: there may not be perfect match between the fusions found by
STAR-Fusion and FusionInspector (since the latter reruns STAR on the
mini contig; also, selfie fusions are ignored).  If no match can be
found, the whole fusion is skipped.
}
