% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fusioninspector.R
\name{import_fusioninspector}
\alias{import_fusioninspector}
\title{read the results from runnning FusionInspector}
\usage{

  import_fusioninspector(filename = "FusionInspector-inspect/finspector.igv.FusionJuncSpan",
  limit)
}
\arguments{
\item{filename}{Which file to import. Note that}
}
\value{
The report is returned as \code{data.frame}
}
\description{
This function is called when \code{\link{import_star_fusion}} is called
with \code{use_fusioninspector=TRUE}. It will override coordinates
(chromosome, strand, and breakpoint) with those corresponding to
FusionInspector's "mini genome". The resulting list of fusion objects
can currently only be plotted with \code{\link{plotFusionReadsSimple}}.
}
\details{
Note: there may not be perfect match between the fusions found by
STAR-Fusion and FusionInspector (since the latter reruns STAR on the
mini contig; also, selfie fusions are ignored).  If no match can be
found, the whole fusion is skipped.
}
