% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fusionmap.R
\name{import_fusionmap}
\alias{import_fusionmap}
\title{Import results from a FusionMap run into a list of Fusion objects.}
\usage{
import_fusionmap(filename, genome_version, limit)
}
\arguments{
\item{filename}{Filename for the FusionMap PairedEndFusionReport.txt results
file.}

\item{genome_version}{Which genome was used in mapping (hg19, hg38, etc.).}

\item{limit}{A limit on how many lines to read.}
}
\value{
A list of Fusion objects.
}
\description{
A function that imports the results from a FusionMap run, typically from
a InputFastq.FusionReport.txt file, into a list of Fusion objects.
}
\examples{
fusionmapData <- system.file(
  "extdata",
  "FusionMap_01_TestDataset_InputFastq.FusionReport.txt",
  package = "chimeraviz")
fusions <- import_fusionmap(fusionmapData, "hg19", 3)
# This should import a list of 3 fusions described in Fusion objects.

}
