% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFusion.R
\name{importFunctionNonUCSC}
\alias{importFunctionNonUCSC}
\title{Alternative import function for Gviz::AlignmentsTrack}
\usage{
importFunctionNonUCSC(file, selection)
}
\arguments{
\item{file}{The bamfile.}

\item{selection}{Which regions to get from the bamfile.}
}
\value{
A GRanges object with coverage data for the selection.
}
\description{
This alternative import function for use with Gviz::AlignmentsTrack imports
a bamfile with non-UCSC chromosome names.
}
