% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStarfusion.R
\name{importFusionInspector}
\alias{importFusionInspector}
\title{read the results from runnning FusionInspector}
\usage{

  importFusionInspector(filename = "FusionInspector-inspect/finspector.igv.FusionJuncSpan",
  limit = Inf)
}
\description{
Note: there may not be perfect match between the fusions found by STAR-Fusion
and FusionInspector (since the latter reruns STAR on the mini contig; also,
selfie fusions are ignored).
If no match can be found, the whole fusion is skipped
}
