% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importInfusion.R
\name{importInfusion}
\alias{importInfusion}
\title{Import results from an InFusion run into a list of Fusion objects.}
\usage{
importInfusion(filename, genomeVersion, limit)
}
\arguments{
\item{filename}{Filename for the jaffa_results.csv file.}

\item{genomeVersion}{Which genome was used in mapping (hg19, hg38, etc.).}

\item{limit}{A limit on how many lines to read.}
}
\value{
A list of Fusion objects.
}
\description{
A function that imports the results from an InFusion run nto a list of Fusion
objects.
}
\examples{
infusionData <- system.file(
  "extdata",
  "infusion_fusions.txt",
  package = "chimeraviz")
fusions <- importInfusion(infusionData, "hg19", 3)
# This should import a list of 3 fusions described in Fusion objects.

}

