% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{partnerGeneJunctionSequence}
\alias{partnerGeneJunctionSequence}
\alias{partnerGeneJunctionSequence,PartnerGene-method}
\title{Get the junction sequence from a PartnerGene object}
\usage{
partnerGeneJunctionSequence(x)

\S4method{partnerGeneJunctionSequence}{PartnerGene}(x)
}
\arguments{
\item{x}{The PartnerGene object you wish to retrieve the junction sequence for.}
}
\value{
The upstream fusion partner gene junction sequence.
}
\description{
This getter retrieves the junction sequence from a PartnerGene object
}
\examples{
# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
fusion <- fusions[[1]]
# Get the junction sequence from the upstream fusion partner gene
partnerGeneJunctionSequence(upstreamPartnerGene(fusion))

}
