% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{plotCircle}
\alias{plotCircle}
\title{Create a circle plot of the given fusions.}
\usage{
plotCircle(fusionList)
}
\arguments{
\item{fusionList}{A list of Fusion objects.}
}
\value{
Creates a circle plot.
}
\description{
This function takes a list of Fusion objects and creates a circle plot
indicating which chromosomes the fusion genes in the list consists of.
}
\details{
Note that only a limited number of gene names can be shown in the circle plot
due to the limited resolution of the plot. RCircos will automatically limit
the number of gene names shown if there are too many. Also note that fusions
involving mitochondrial DNA will not be shown in this plot.
}
\examples{
defuse833ke <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuse833ke, "hg19", 3)
# Temporary file to store the plot
pngFilename <- tempfile(
  pattern = "circlePlot",
  fileext = ".png",
  tmpdir = tempdir())
# Open device
png(pngFilename, width = 1000, height = 750)
# Plot!
plotCircle(fusions)
# Close device
dev.off()

}
