% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFusionReadsSimple.R
\name{plotContigReads}
\alias{plotContigReads}
\title{Plot fusion with separate junction and spanning reads}
\usage{
plotContigReads(fusion, bam.split, bam.spanning, leftFlank = 1000,
  rightFlank = leftFlank)
}
\arguments{
\item{fusion}{Fusion object}

\item{bam.split}{name of bam file with junction reads in mini contig coords}

\item{bam.spanning}{name of bam file with spanning reads in mini contig coords}

\item{leftFlank, rightFlank}{how much to show left(right) of the split}
}
\description{
Plot fusion with separate junction and spanning reads
}
