% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFusionReads.R
\name{plotFusionReads}
\alias{plotFusionReads}
\title{Create a plot of the reads supporting the given fusion.}
\usage{
plotFusionReads(fusion, showAllNucleotides = TRUE, nucleotideAmount = 10)
}
\arguments{
\item{fusion}{The Fusion object to plot.}

\item{showAllNucleotides}{Boolean indicating whether or not to show all
nucleotides. If FALSE, then only nucleotideAmount amount of nucleotides will
be shown on each end of the fusion junction. If TRUE, then the whole fusion
junction sequence will be shown.}

\item{nucleotideAmount}{The number of nucleotides to show on each end of the
fusion junction sequence. Defaults to 10. Only applicable if
showAllNucleotides is set to TRUE.}
}
\value{
Creates a fusion reads plot.
}
\description{
This function takes a Fusion object and plots the reads supporting the fusion
on top of the fusion sequence (fusion@junctionSequence), provided that
addFusionReadsAlignment() has been run earlier in order to add fusion reads
alignment data to the fusion object.
}
\details{
Note that the package used for plotting, Gviz, is strict on chromosome names.
If the plot produced doesn't show the reads, the problem might be solved by
naming the fusion sequence "chrNA".
}
\examples{
# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
# Find the specific fusion we have aligned reads for
fusion <- getFusionById(fusions, 5267)
bamfile <- system.file(
  "extdata",
  "5267readsAligned.bam",
  package="chimeraviz")
# Add the bam file of aligned fusion reads to the fusion object
fusion <- addFusionReadsAlignment(fusion, bamfile)
# Temporary file to store the plot
pngFilename <- tempfile(
  pattern = "fusionPlot",
  fileext = ".png",
  tmpdir = tempdir())
# Calculate image size based on supporting reads and lenght of junction
# sequence.
imageWidth <- (nchar(partnerGeneJunctionSequence(upstreamPartnerGene(fusion))) +
  nchar(partnerGeneJunctionSequence(downstreamPartnerGene(fusion)))) * 15
imageHeight <- (fusionSplitReadsCount(fusion)+fusionSpanningReadsCount(fusion)) * 20
# Open device
png(pngFilename, width = imageWidth, height = imageHeight)
# Now we can plot
plotFusionReads(fusion)
# Close device
dev.off()

}
\seealso{
addFusionReadsAlignment
}

