% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFusionReadsSimple.R
\name{plotFusionReadsSimple}
\alias{plotFusionReadsSimple}
\title{Simple version of plotFusionReads}
\usage{
plotFusionReadsSimple(fusion, leftFlank = 10000, rightFlank = leftFlank)
}
\description{
Simple version of plotFusionReads
}
