% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_aeron}
\alias{raw_aeron}
\title{Aeron fusion data}
\description{
Documentation for the Aeron example data.
}
\section{aeron_fusion_support.txt}{


This is example data for Aeron provided on GitHub here:
https://github.com/SchulzLab/Aeron/issues/5#issuecomment-603447735
It was downloaded March 25th 2020. The original filename was
aeron_fusion_support_301018_K562_mRNA_PCS109_HumanTranscriptsReduced_HumanUpdated38V5.txt
}

