% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_cytobandhg19}
\alias{raw_cytobandhg19}
\title{Cytoband information HG19}
\description{
Cytoband information for the HG19 assembly from UCSC. Downloaded from
http://hgdownload.cse.ucsc.edu/goldenpath/hg19/database/
}
\section{UCSC.HG19.Human.CytoBandIdeogram.txt}{


This data is used with RCircos in plotCircle().
}

