% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_cytobandhg38}
\alias{raw_cytobandhg38}
\title{Cytoband information HG138}
\description{
Cytoband information for the HG38 assembly from UCSC. Downloaded from
http://hgdownload.cse.ucsc.edu/goldenpath/hg38database/
}
\details{
All _alt or _random entries has been manually removed, as has the chrM entry.
}
\section{UCSC.HG38.Human.CytoBandIdeogram.txt}{


This data is used with RCircos in plotCircle().
}

