% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_defuse}
\alias{raw_defuse}
\title{deFuse data}
\description{
Documentation for the deFuse example data.
}
\section{defuse_833ke_results.filtered.tsv}{


This file has the results from a run of deFuse-0.7.0 on the 833ke cell line.
The program was ran with the standard configuration, but with the parameter
span_count_threshold=5 instead of the standard 3. The resulting
results.filtered.tsv file was then manually filtered to only include 17
fusion events in the interest of saving computing time for tests and
examples. The original results contained 171 fusion events.
}

\section{reads_supporting_defuse_fusion_5267.*.fq}{


These two files, reads_supporting_defuse_fusion_5267.1.fq and
reads_supporting_defuse_fusion_5267.2.fq, contains the reads that support the
fusion event with cluster_id 5267.
}

\section{5267readsAligned.bam}{


The bamfile 5267readsAligned.bam and the 5267readsAligned.bam.bai index file
contains the reads supporting the fusion event with cluster_id 5267 aligned
to the fusion sequence. It is used with plot_fusion_reads().
}

