% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_ericscript}
\alias{raw_ericscript}
\title{EricScript data}
\description{
Documentation for the EricScript example data.
}
\section{ericscript_SRR1657556.results.total.tsv}{


This is example data thankfully provided by EricScript author Matteo Benelli.
}

