% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_fusion5267proteindomains}
\alias{raw_fusion5267proteindomains}
\title{protein_domains_5267 bed file}
\description{
Documentation for the protein_domains_5267.bed file containing protein
domains for the genes in the fusion with cluster_id=5267.
}
\section{protein_domains_5267.bed}{


This file is an excerpt from a larger file that we created by:
- downloading domain name annotation from Pfam database (PfamA version 31)
  and domain region annotation from Ensembl database through BioMart API
- switching the domain coordinates in the protein level to these in
  transcript level.
}

