% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_fusion5267reads}
\alias{raw_fusion5267reads}
\title{Fusion5267and11759 bamfile}
\description{
Documentation for the fusion5267and11759reads.bam file containing reads
mapped to the region where the genes in the fusions with cluster_id=5267 and
cluster_id=11759 is.
}
\section{fusion5267and11759reads.bam}{


This file is the result of running these commands:

samtools view -b original_bamfile.bam "1:28831455-28866812" "1:109189912-109205148" "12:8608225-8677832" > fusion5267and11759reads.bam
samtools index fusion5267and11759reads.bam fusion5267and11759reads.bam.bai

where we extract the reads mapping to the region where we know the fusions
with cluster_id=5267 and cluster_id=11759 from the deFuse example data is.

The original_bamfile.bam is from a study of the 833KE cell line by Andreas M. Hoff et al., documented in the paper [Identification of Novel Fusion Genes in Testicular Germ Cell Tumors](http://cancerres.aacrjournals.org/content/76/1/108.full).
}

