% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_fusion5267readsBedGraph}
\alias{raw_fusion5267readsBedGraph}
\title{Fusion5267and11759 bedGraph file}
\description{
Documentation for the fusion5267and11759reads.bedGraph file containing read
count data from the regions of the fusion event with cluster_id=5267.
}
\section{fusion5267and11759reads.bedGraph}{


This file is the result of running this command:

bedtools genomecov -ibam fusion5267and11759reads.bam -bg > fusion5267and11759reads.bam.bedGraph

fusion5267and11759reads.bam has its own documentation entry for how it was
created.
}

