% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_fusioncatcher}
\alias{raw_fusioncatcher}
\title{Fusioncatcher data}
\description{
Documentation for the Fusioncatcher example data.
}
\section{fusioncatcher_833ke_final-list-candidate-fusion-genes.txt}{


This file has the results from a run of Fusioncatcher-0.99.3e on the 833ke
cell line. The program was ran with the standard configuration file and with
the parameters "-p 8 -z --keep-preliminary".
}

