% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_fusionmap}
\alias{raw_fusionmap}
\title{FusionMap data}
\description{
Documentation for the FusionMap example data.
}
\section{FusionMap_01_TestDataset_InputFastq.FusionReport.txt}{


This is example data provided with the FusionMap version released 2015-03-31.
}

