% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_Homo_sapiens.GRCh37.74}
\alias{raw_Homo_sapiens.GRCh37.74}
\title{Homo_sapiens.GRCh37.74_subset.gtf}
\description{
Homo_sapiens.GRCh37.74_subset.gtf
}
\section{Homo_sapiens.GRCh37.74_subset.gtf}{


The Homo_sapiens.GRCh37.74.gtf file is a subset version of the Ensembl
Homo_sapiens.GRCh37.74.gtf file, located here:
\url{ftp://ftp.ensembl.org/pub/release-74/gtf/homo_sapiens}. This gtf file
contains transcripts for the partner genes in two of the fusion transcripts
from the deFuse example data provided with this package: The fusion
transcript with cluster_id=5267, and the fusion transcript with
cluster_id=11759.

The file is the result of running this command:

# grep "ENST00000373831\|ENST00000373832\|ENST00000373833\|ENST00000398958\|ENST00000411533\|ENST00000419074\|ENST00000427469\|ENST00000429051\|ENST00000430407\|ENST00000434290\|ENST00000478232\|ENST00000486790\|ENST00000370031\|ENST00000370032\|ENST00000402983\|ENST00000420055\|ENST00000483729\|ENST00000493676\|ENST00000382073\|ENST00000299665\|ENST00000382064" Homo_sapiens.GRCh37.74.gtf > Homo_sapiens.GRCh37.74_subset.gtf

The transcript names given in the command above are all transcripts available
for the genes CLEC6A, CLEC4D, HENMT1, and RCC1 in Ensembl version 74.
}

\section{Homo_sapiens.GRCh37.74.sqlite}{


The Homo_sapiens.GRCh37.74.sqlite file is the sqlite database that the
Ensembldb package creates from the corresponding gtf file. It was created
using this command:

# ensDbFromGtf(
#   gtf = "Homo_sapiens.GRCh37.74_subset.gtf",
#   organism = "Homo_sapiens",
#   genomeVersion = "GRCh37",
#   version = 74)
}

