% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_infusion}
\alias{raw_infusion}
\title{InFusion data}
\description{
Documentation for the InFusion example data.
}
\section{infusion_fusions.txt}{


This is example data from the InFusion getting started page located at
https://bitbucket.org/kokonech/infusion/wiki/Getting%20Started .
}

