% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_jaffa}
\alias{raw_jaffa}
\title{JAFFA data}
\description{
Documentation for the JAFFA example data.
}
\section{jaffa_results.csv}{


This is example data from the described JAFFA example run documented at
https://github.com/Oshlack/JAFFA/wiki/Example .
}

