% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_oncofuse}
\alias{raw_oncofuse}
\title{oncofuse data}
\description{
Documentation for the oncofuse example data.
}
\section{oncofuse.outfile}{


The example output from oncofuse was kindly provided by Lavinia G here:
https://github.com/stianlagstad/chimeraviz/issues/47#issuecomment-409773158
}

