% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_prada}
\alias{raw_prada}
\title{PRADA data}
\description{
Documentation for the PRADA example data.
}
\section{PRADA.acc.fusion.fq.TAF.tsv}{


This is example data thankfully provided by PRADA authors Siyuan Zheng and
Roeland Verhaak.
}

