% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_soapfuse}
\alias{raw_soapfuse}
\title{SOAPfuse data}
\description{
Documentation for the SOAPfuse example data.
}
\section{soapfuse_833ke_final.Fusion.specific.for.genes}{


This file has the results from a run of soapfuse-1.26 on the 833ke cell line.
The program was ran with the standard configuration file.
}

