% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_starfusion}
\alias{raw_starfusion}
\title{STAR-Fusion data}
\description{
Documentation for the STAR-Fusion example data.
}
\section{star-fusion.fusion_candidates.final.abridged.txt}{


This example data was retrieved from the STAR-Fusion github page June 2.nd
2017.
}

