% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,Fusion-method}
\alias{show,Fusion-method}
\title{Show method for the Fusion class.}
\usage{
\S4method{show}{Fusion}(object)
}
\arguments{
\item{object}{A Fusion object}
}
\value{
Shows information about a Fusion object.
}
\description{
Show method for the Fusion class.
}
