% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{show,PartnerGene-method}
\alias{show,PartnerGene-method}
\title{Show method for the PartnerGene class.}
\usage{
\S4method{show}{PartnerGene}(object)
}
\arguments{
\item{object}{A PartnerGene object}
}
\value{
Shows information about a PartnerGene object.
}
\description{
Show method for the PartnerGene class.
}
