% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{upstreamPartnerGene}
\alias{upstreamPartnerGene}
\alias{upstreamPartnerGene,Fusion-method}
\alias{upstreamPartnerGene<-}
\alias{upstreamPartnerGene<-,Fusion-method}
\title{Get the upstream fusion partner gene}
\usage{
upstreamPartnerGene(x)

\S4method{upstreamPartnerGene}{Fusion}(x)

upstreamPartnerGene(object) <- value

\S4method{upstreamPartnerGene}{Fusion}(object) <- value
}
\arguments{
\item{x}{The Fusion object you wish to retrieve the upstream PartnerGene object for.}

\item{object}{The Fusion object you wish to set a new upstream PartnerGene object for.}

\item{value}{The new PartnerGene object.}
}
\value{
The upstream PartnerGene object.
}
\description{
This getter retrieves the upstream PartnerGene object.

This sets the upstream PartnerGene object of a Fusion object
}
\examples{
# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
fusion <- fusions[[1]]
# Get the upstream fusion partner gene
upstreamPartnerGene(fusion)

# Load data
defuseData <- system.file(
  "extdata",
  "defuse_833ke_results.filtered.tsv",
  package="chimeraviz")
fusions <- importDefuse(defuseData, "hg19", 1)
fusion <- fusions[[1]]
# Set the upstream PartnerGene object to be the same as the downstream PartnerGene object
upstreamPartnerGene(fusion) <- downstreamPartnerGene(fusion)

}
