% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{chimeraviz-internals-scaleListToInterval}
\alias{.scaleListToInterval}
\alias{chimeraviz-internals-scaleListToInterval}
\title{Scale a vector of numeric values to an interval.}
\usage{
.scaleListToInterval(theList, newMin, newMax)
}
\arguments{
\item{theList}{A vector of numeric values.}

\item{newMin}{Minimum value for the new interval.}

\item{newMax}{Maximum value for the new interval.}
}
\value{
A data frame with fusion link data compatible with RCircos::RCircos.Link.Plot()

# @examples # Apparently examples shouldn't be set on private functions
list012 <- c(0,1,2)
.scaleListToInterval(list012, 1, 3)
# [1] 1 2 3
}
\description{
This function takes a vector of numeric values as well as an interval
[newMin, newMax] that the numeric values will be scaled (normalized) to.
}

