% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStarfusion.R
\name{importFusionInspector}
\alias{importFusionInspector}
\title{read the results from runnning FusionInspector}
\usage{

  importFusionInspector(filename = "FusionInspector-inspect/finspector.igv.FusionJuncSpan",
  limit = Inf)
}
\description{
read the results from runnning FusionInspector
}
