% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{raw_chimpipe}
\alias{raw_chimpipe}
\title{ChimPipe data}
\description{
Documentation for the ChimPipe example data.
}
\section{chimericJunctions_MCF-7.txt}{


This is example data from the ChimPipe tutorial all-in-one package located at
https://chimpipe.readthedocs.io/en/latest/tutorial.html#download-all-in-one-package
It was downloaded March 10th 2020
}

